commit after lunch, only little changes

+++ project-01/shuffla/db/migrate/20140701000126_add_points_to_questions.rb
@@ -1,5 +1,5 @@
 class AddPointsToQuestions < ActiveRecord::Migration
   def change
-    add_column :questions, :points, :decimal
+    add_column :questions, :points, :float
   end
 end
